Address warnings

+++ src/libfoenix/uart.c
@@ -48,7 +48,7 @@ uint8_t * uart_get_base(int16_t uart) {
  * 
  * @return the value that can be used with _set_bps, or 0 if invalid argument
  */
-uint32_t uart_baud_to_code(uint16_t i) {
+uint32_t uart_baud_to_code(uint32_t i) {
     switch (i) {
     case 300:  return UART_300;
     case 1200: return UART_1200;
